Last Structure
I understood that if i make the width or the height equal to 100% it wont work perfectly in every Monitor (resolution) so i decided to use Px for my divs from now on, i also learned that you have to define everything (margins, width and height) to it so it works better, also make the perfect divs from the beginning, in the end i got better. it was worth it.